basket auto and code cleanup
